+++ RELEASE_NOTES.rst
@@ -9,8 +9,11 @@ Release Notes
 
 
 
-.. Upcoming Release
-.. ================
+Upcoming Release
+================
+
+* The functions ``add_raster()`` and ``add_geometry()`` of the ``ExclusionContainer`` can now directly read from URL. 
+
 
 Version 0.2.11
 ==============